New "fcontact" model class
